+++ src/librawrtc/candidate_helper.h
@@ -11,6 +11,16 @@ struct rawrtc_candidate_helper_stun_session {
 };
 
 /*
+ * TURN client session.
+ */
+struct rawrtc_candidate_helper_turn_session {
+    struct le le;
+    struct rawrtc_candidate_helper* candidate_helper;
+    struct turnc* turn_client;
+    struct rawrtc_ice_server_url* url;
+};
+
+/*
  * Local candidate helper.
  */
 struct rawrtc_candidate_helper {
@@ -21,6 +31,7 @@ struct rawrtc_candidate_helper {
     uint_fast8_t srflx_pending_count;
     struct list stun_sessions;
     uint_fast8_t relay_pending_count;
+    struct list turn_sessions;
 };
 
 enum rawrtc_code rawrtc_candidate_helper_create(
@@ -52,6 +63,17 @@ enum rawrtc_code rawrtc_candidate_helper_stun_session_add(
     struct rawrtc_candidate_helper_stun_session* const session,
     struct rawrtc_candidate_helper* const candidate_helper,
     struct stun_keepalive* const stun_keepalive
+);
+
+enum rawrtc_code rawrtc_candidate_helper_turn_session_create(
+    struct rawrtc_candidate_helper_turn_session** const sessionp, // de-referenced
+    struct rawrtc_ice_server_url* const url
+);
+
+enum rawrtc_code rawrtc_candidate_helper_turn_session_add(
+    struct rawrtc_candidate_helper_turn_session* const session,
+    struct rawrtc_candidate_helper* const candidate_helper,
+    struct turnc* const turn_client
 );
 
 bool rawrtc_candidate_helper_remove_stun_sessions_handler(
